% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eco.utils.indices-package.R
\docType{package}
\name{Eco.utils.indices-package}
\alias{Eco.utils.indices}
\alias{Eco.utils.indices-package}
\title{Eco.utils.indices: Ecological Indicators}
\description{
Tidyverse wrapper for biomonitoR functions.
}
\author{
\strong{Maintainer}: Paul Palmer \email{p.j.palmert@lboro.ac.uk} (\href{https://orcid.org/0000-0002-1674-3728}{ORCID})

}
\keyword{internal}
