% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_bmwp.R
\name{tidy_bmwp}
\alias{tidy_bmwp}
\title{Tidy BMWP}
\usage{
tidy_bmwp(x, aspt = TRUE, ...)
}
\arguments{
\item{x}{column of taxon names followed by one or more columns of
observations as number of individuals.}

\item{aspt}{Average Score Per Taxon}

\item{...}{Other parameters}
}
\value{
dataframe of values
}
\description{
Calculate BMWP index
}
\examples{

data(macro_ex)




}
