% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_coleoptera.R
\name{tidy_coleoptera}
\alias{tidy_coleoptera}
\title{Count Coleoptera Families}
\usage{
tidy_coleoptera(x)
}
\arguments{
\item{x}{data}
}
\value{
dataframe of coleoptera
}
\description{
Count the number of coleoptera families in the data
}
