% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_dragonfly_allies.R
\name{tidy_dragonfly_allies}
\alias{tidy_dragonfly_allies}
\title{Dragonflies and Allies}
\usage{
tidy_dragonfly_allies(x)
}
\arguments{
\item{x}{dataframe of data.}
}
\value{
dataframe containing count of dragonflies and allies
}
\description{
Count of dragonflies, alderflies, mayflies and damselflies.
}
