% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_indicators.R
\name{tidy_indicators}
\alias{tidy_indicators}
\title{Tidyverse Get All Indicators}
\usage{
tidy_indicators(x)
}
\arguments{
\item{x}{The data}
}
\value{
A dataframe of all the indicators
}
\description{
A tidyvwerse form of calculating all the indicators in one go.
}
