% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_observation_data.R
\name{tidy_observation_data}
\alias{tidy_observation_data}
\title{Tidy Observation Data}
\usage{
tidy_observation_data(x)
}
\arguments{
\item{x}{data frame Must contain column Taxa (not case) with named observation columns to the right}
}
\value{
data frame with Taxa and Observation columns
}
\description{
Cleans observation data for use with biomonitoR
}
