% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_observation_maker.R
\name{tidy_observation_maker}
\alias{tidy_observation_maker}
\title{Tidy Function To Make Observation Data}
\usage{
tidy_observation_maker(meta.data.list, observation.list)
}
\arguments{
\item{meta.data.list}{Named list that must contain recordID}

\item{observation.list}{Named list of families and numbers.}
}
\value{
dataframe of observations
}
\description{
Requires two named lists to work
}
\examples{

my.meta.data <- list(recordID = "my.records")
my.observations <- list(Dryopidae = 1)

tidy_observation_maker(my.meta.data, my.observations)



}
