% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_test_names.R
\name{tidy_test_names}
\alias{tidy_test_names}
\title{Title}
\usage{
tidy_test_names(x)
}
\arguments{
\item{x}{data}
}
\value{
list comprising problem.logic True or False. suggested.taxa as a
dataframe of suggested taxa.
}
\description{
Title
}
