% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_whpt.R
\name{tidy_whpt}
\alias{tidy_whpt}
\title{Tidy WHPT}
\usage{
tidy_whpt(x, aspt = TRUE, ...)
}
\arguments{
\item{x}{column of taxon names followed by one or more columns of
observations as number of individuals.}

\item{aspt}{Average Score Per Taxon}

\item{...}{Other parameters}
}
\value{
dataframe of values
}
\description{
Calculate WHPT index
}
